+++ src/allegrord.cpp
@@ -34,7 +34,7 @@ public:
     long parse_int(string &field);
     size_t find_real_in(string &field, size_t n);
     double parse_real(string &field);
-    void parse_error(string &field, long offset, const char *message);
+    void parse_error(string &field, size_t offset, const char *message);
     double parse_dur(string &field, double base);
     double parse_after_dur(double dur, string &field, int n, double base);
     double parse_loud(string &field);
@@ -51,7 +51,7 @@ public:
 double Alg_reader::parse_pitch(string &field)
 {
     if (isdigit(field[1])) {
-        int last = find_real_in(field, 1);
+        size_t last = find_real_in(field, 1);
         string real_string = field.substr(1, last - 1);
         return atof(real_string.c_str());
     } else {
@@ -495,9 +495,9 @@ size_t Alg_reader::find_real_in(string &field, size_t n)
 double Alg_reader::parse_real(string &field)
 {
     const char *msg = "Real expected";
-    int last = find_real_in(field, 1);
+    size_t last = find_real_in(field, 1);
     string real_string = field.substr(1, last - 1);
-    if (last <= 1 || last < static_cast<int>(field.length())) {
+    if (last <= 1 || last < field.length()) {
        parse_error(field, 1, msg);
        return 0;
     }
@@ -505,12 +505,12 @@ double Alg_reader::parse_real(string &field)
 }
 
 
-void Alg_reader::parse_error(string &field, long offset, const char *message)
+void Alg_reader::parse_error(string &field, size_t offset, const char *message)
 {
-    int position = line_parser.pos - field.length() + offset;
+    size_t position = line_parser.pos - field.length() + offset;
     error_flag = true;
     puts(line_parser.str->c_str());
-    for (int i = 0; i < position; i++) {
+    for (size_t i = 0; i < position; i++) {
         putc(' ', stdout);
     }
     putc('^', stdout);
@@ -526,7 +526,7 @@ double Alg_reader::parse_dur(string &field, double base)
     const char *msg = "Duration expected";
     const char *durs = "SIQHW";
     const char *p;
-    int last;
+    size_t last;
     double dur;
     if (field.length() < 2) {
         // fall through to error message
@@ -565,7 +565,7 @@ double Alg_reader::parse_after_dur(double dur, string &field,
         return parse_after_dur(dur * 1.5, field, n + 1, base);
     }
     if (isdigit(field[n])) {
-        int last = find_real_in(field, n);
+        size_t last = find_real_in(field, n);
         string a_string = field.substr(n, last - n);
         double f = atof(a_string.c_str());
         return parse_after_dur(dur * f, field, last, base);
